fix: keep kobo in analytics totals instead of integer-truncating
Sequelize Model.sum() parsed INTEGER money columns with parseInt, so
fractional amounts disagreed with list-page totals.

+++ app/models/expense.ts
@@ -22,7 +22,7 @@ export default sequelize.define('expense', {
     primaryKey: true,
   },
   date: Sequelize.DATE,
-  amount: Sequelize.INTEGER,
+  amount: Sequelize.DOUBLE,
   type: Sequelize.STRING,
   note: Sequelize.STRING,
   postedBy: Sequelize.STRING,
